requirements.txt
setuptools
django
erniebot
paddlepaddle
